Client Rework
